Refactoring/cleanup: CSV crosswalk now in a separate utility function
