+++ src/pages/[lang]/projects/[project]/export/csv.ts
@@ -1,16 +1,9 @@
-import Papa from 'papaparse';
 import { getAllDocumentLayersInProject, getAllLayersInProject, getProjectPolicies } from '@backend/helpers';
 import { getAnnotations } from '@backend/helpers/annotationHelpers';
 import { getMyProfile } from '@backend/crud';
 import { createSupabaseServerClient } from '@backend/supabaseServerClient';
 import type { APIRoute } from 'astro';
-import { Visibility, type SupabaseAnnotationBody, type SupabaseAnnotationTarget } from '@recogito/annotorious-supabase';
-
-const getComments = (bodies: SupabaseAnnotationBody[]) =>
-  bodies.filter(b => b.purpose === 'commenting').map(b => b.value);
-
-const getTags = (bodies: SupabaseAnnotationBody[]) =>
-  bodies.filter(b => b.purpose === 'tagging').map(b => b.value);
+import { annotationsToCSV } from 'src/util/export/csv';
 
 export const get: APIRoute = async ({ params, request, cookies, url }) => {
   const supabase = await createSupabaseServerClient(request, cookies);
@@ -60,30 +53,11 @@ export const get: APIRoute = async ({ params, request, cookies, url }) => {
       JSON.stringify({ message: 'Error retrieving annotations' }), 
       { status: 500 }); 
   }
-  
-  const findDocument = (layerId: string) =>
-    layers.data.find(l => l.id === layerId)?.document?.name;
 
-  const getLastUpdated = (target: SupabaseAnnotationTarget, bodies: SupabaseAnnotationBody[]) => {
-    const sorted = [target, ...bodies];
-    sorted.sort((a, b) => a.updated && b.updated ? a.updated > b.updated ? -1 : 1 : 0);
-    return sorted[0].updated;
-  }
-
-  const csv = annotations.data.map(a => ({
-    id: a.id,
-    document: findDocument(a.layer_id!)!,
-    created: a.target.created,
-    updated: getLastUpdated(a.target, a.bodies),
-    comments: getComments(a.bodies).join('|'),
-    tags: getTags(a.bodies).join('|'),
-    is_private: a.visibility === Visibility.PRIVATE
-  }));
-
-  csv.sort((a, b) => a.document > b.document ? -1 : 1);
+  const csv = annotationsToCSV(annotations.data, layers.data);
 
   return new Response(    
-    Papa.unparse(csv),
+    csv,
     { 
       headers: { 
         'Content-Type': 'text/csv',
